fix(widgets): extract advanceNextUpQueue to utility; wire test to production import
